+++ src/algorithms/constructors/mod.rs
@@ -31,4 +31,5 @@ pub mod star;
 pub mod symmetric_tree;
 pub mod tree_from_parent_vector;
 pub mod triangular_lattice;
+pub mod turan;
 pub mod wheel;